docs: add documentation to new calibration module
