+++ docs/package/calibration.md
@@ -1,0 +1,8 @@
+::: pychilaslasers.calibration
+    options:
+        summary: true        
+    filters:
+        - "!^__"                     # hide dunders
+        - "!^_"                      # hide private names
+
+
